requirements.txt
web3
pandas
streamlit
plotly
psycopg2-binary
python-dotenv
requests
schedule
